mini-todo-backend/pyproject.toml
[project]
name = "mini-todo-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi>=0.115.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.19",
    "ruff>=0.8.1",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.32.1",
]

[tool.ruff]
line-length = 120


